ARM: socfpga: Add SPL fitImage FPGA bitstream loading
Add a hook implementation for loading the FPGA bitstream embedded
in the fitImage. The FPGA image has to be loaded in smaller chunks
as the DRAM is not available just yet. Using the fitImage mechanism
means that loading the bitstream and loading the U-Boot image use
exactly the same code.

Signed-off-by: Marek Vasut <marex@denx.de>
Cc: Chin Liang See <chin.liang.see@intel.com>
Cc: Dinh Nguyen <dinguyen@kernel.org>